refactor: 공통 메서드의 테스트는 Piece 테스트로 이동
